+++ src/components/profile/UserListDialog.js
@@ -1,14 +1,14 @@
 import React from "react";
 import Dialog from "@mui/material/Dialog";
 import DialogTitle from "@mui/material/DialogTitle";
-import SuggestedProfile from "./SuggestedProfile";
+import FollowProfileLink from "./FollowProfileLink";
 
 function UserListDialog({ open, handleClose, userList, title }) {
   return (
     <Dialog onClose={handleClose} open={open}>
       <DialogTitle>{title}</DialogTitle>
       {userList.map((user) => (
-        <SuggestedProfile
+        <FollowProfileLink
           key={user.id}
           userId={user.id}
           username={user.username}
